rename inconsistent packages from "fourier" to "fouriermethod"
