+++ src/main/java/net/finmath/fouriermethod/calibration/models/calibrationDiagram.png
